rpk: move `cloud secrets` -> `secret`
This has not been released to a stable version
yet, so we are making this change without leaving
a backcompat command.
