+++ src/go/rpk/pkg/cli/cloud/BUILD
@@ -14,7 +14,6 @@ go_library(
         "//src/go/rpk/pkg/cli/cloud/byoc",
         "//src/go/rpk/pkg/cli/cloud/cluster",
         "//src/go/rpk/pkg/cli/cloud/resourcegroup",
-        "//src/go/rpk/pkg/cli/cloud/security",
         "//src/go/rpk/pkg/cli/container/common",
         "//src/go/rpk/pkg/cli/profile",
         "//src/go/rpk/pkg/config",
