+++ src/go/rpk/pkg/cli/cloud/cloud.go
@@ -14,7 +14,6 @@ import (
 	"github.com/redpanda-data/redpanda/src/go/rpk/pkg/cli/cloud/byoc"
 	"github.com/redpanda-data/redpanda/src/go/rpk/pkg/cli/cloud/cluster"
 	"github.com/redpanda-data/redpanda/src/go/rpk/pkg/cli/cloud/resourcegroup"
-	"github.com/redpanda-data/redpanda/src/go/rpk/pkg/cli/cloud/security"
 	"github.com/redpanda-data/redpanda/src/go/rpk/pkg/config"
 	"github.com/spf13/afero"
 	"github.com/spf13/cobra"
@@ -31,7 +30,6 @@ func NewCommand(fs afero.Fs, p *config.Params, execFn func(string, []string) err
 		byoc.NewCommand(fs, p, execFn),
 		cluster.NewCommand(fs, p),
 		resourcegroup.NewCommand(fs, p),
-		security.NewCommand(fs, p),
 		newLoginCommand(fs, p),
 		newLogoutCommand(fs, p),
 	)
