+++ src/go/rpk/pkg/cli/security/secret/list.go
