+++ src/go/rpk/pkg/cli/security/secret/secret_test.go
