+++ src/go/rpk/pkg/cli/security/secret/update.go
@@ -11,6 +11,7 @@ package secret
 
 import (
 	"fmt"
+	"strings"
 
 	dataplanev1 "buf.build/gen/go/redpandadata/dataplane/protocolbuffers/go/redpanda/api/dataplane/v1"
 	"connectrpc.com/connect"
@@ -21,15 +22,17 @@ import (
 	"github.com/spf13/cobra"
 )
 
-func newDeleteCommand(fs afero.Fs, p *config.Params) *cobra.Command {
-	var secretName string
+func newUpdateCommand(fs afero.Fs, p *config.Params) *cobra.Command {
+	var secretName, secretValue string
 
 	cmd := &cobra.Command{
-		Use:   "delete",
-		Short: "Delete an existing secret",
-		Long:  "Delete an existing secret from your Redpanda Cloud cluster",
+		Use:   "update",
+		Short: "Update an existing secret",
+		Long:  "Update an existing secret for your Redpanda Cloud cluster",
 		Run: func(cmd *cobra.Command, _ []string) {
-			out.MaybeDie(validateSecretName(secretName), "invalid secret name: %v")
+			err := validateSecretName(secretName)
+			out.MaybeDie(err, "invalid secret name: %v", err)
+
 			p, err := p.LoadVirtualProfile(fs)
 			out.MaybeDie(err, "rpk unable to load config: %v", err)
 			if !p.FromCloud {
@@ -48,18 +51,21 @@ func newDeleteCommand(fs afero.Fs, p *config.Params) *cobra.Command {
 			cl, err := publicapi.NewDataPlaneClientSet(url, p.CurrentAuth().AuthToken)
 			out.MaybeDie(err, "unable to initialize cloud client: %v", err)
 
-			request := &dataplanev1.DeleteSecretRequest{
-				Id: secretName,
+			request := &dataplanev1.UpdateSecretRequest{
+				Id:         strings.ToUpper(secretName),
+				SecretData: []byte(secretValue),
+				Scopes:     []dataplanev1.Scope{dataplanev1.Scope_SCOPE_REDPANDA_CONNECT},
 			}
-			_, err = cl.Secrets.DeleteSecret(cmd.Context(), &connect.Request[dataplanev1.DeleteSecretRequest]{Msg: request})
-			out.MaybeDie(err, "unable to delete secret: %v", err)
-
-			fmt.Printf("Secret %s deleted successfully \n", secretName)
+			response, err := cl.Secrets.UpdateSecret(cmd.Context(), connect.NewRequest(request))
+			out.MaybeDie(err, "unable to update secret: %v", err)
+			fmt.Printf("Secret %s updated successfully \n", response.Msg.Secret.Id)
 		},
 	}
 
-	cmd.Flags().StringVar(&secretName, "name", "", "Name of the secret to delete, must be uppercase and can only contain letters, digits, and underscores")
+	cmd.Flags().StringVar(&secretName, "name", "", "Name of the secret, must be uppercase and can only contain letters, digits, and underscores")
+	cmd.Flags().StringVar(&secretValue, "value", "", "New secret value of the secret")
 	cmd.MarkFlagRequired("name")
+	cmd.MarkFlagRequired("value")
 
 	return cmd
 }
